+++ src/lib.rs
@@ -1,1 +1,5 @@
 pub mod board_generator;
+pub mod events;
+pub mod game;
+pub mod screen_renderer;
+pub mod sdl_context;
